+++ src/main/java/racingcar/service/RacingCarGameService.java
@@ -1,0 +1,4 @@
+package racingcar.service;
+
+public interface RacingCarGameService {
+}
